Add files via upload
Trying to save names in an arrayList, but so far lack the knowledge to add additional functionality and how to switch between players and save their progress. I do know that there need to be separate classes for player, dice and different events but still difficult to grasp the concepts.